+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - EndShipper.update()
   - CarrierAccount.update()
 - [FIXED] Typo, `ServiceUnavailablError` is now `ServiceUnavailableError`
+- [FIXED] Typo: `Smartrate` is now `SmartRate`, `apikeys` is now `apiKey`, `scanform` is now `scanForm`
 
 ## v6.0.0-rc1 (2022-11-30)